Add files via upload
This is the code for a second bot that only polls r/PrequelMemes rather instead of r/all.  This allows it to catch all the comments on r/PrequelMemes as reddit's rate limiting was causing the bot to miss a lot.  This gets the comments and then sends a POST to the main server.  It is then checked against a mySQL database with all of the comments to make sure it hasn't been replied to already.